requirements.txt
azure-storage-blob==12.19.1
python-dotenv==1.0.1
